requirements.txt
gradio>=4.0.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
piexif>=1.1.3
scipy>=1.9.0
pathlib 